better naming of PhraseApp related stuff
