+++ app/Extensions/PhraseAppTranslator.php
@@ -22,42 +22,24 @@
  * SOFTWARE.
  */
 
-namespace Pterodactyl\Providers;
+namespace Pterodactyl\Extensions;
 
-use Pterodactyl\Extensions\Translator as PterodactylTranslator;
-use Illuminate\Translation\Translator;
-use Illuminate\Translation\TranslationServiceProvider;
+use Illuminate\Translation\Translator as LaravelTranslator;
 
-class TranslationProvider extends TranslationServiceProvider {
-
+class PhraseAppTranslator extends LaravelTranslator
+{
     /**
-     * Register the service provider.
+     * Get the translation for the given key.
      *
-     * @return void
+     * @param  string  $key
+     * @param  array   $replace
+     * @param  string|null  $locale
+     * @param  bool  $fallback
+     * @return string|array|null
      */
-    public function register()
+    public function get($key, array $replace = [], $locale = null, $fallback = true)
     {
-        $this->registerLoader();
-
-        $this->app->singleton('translator', function ($app) {
-            $loader = $app['translation.loader'];
-
-            // When registering the translator component, we'll need to set the default
-            // locale as well as the fallback locale. So, we'll grab the application
-            // configuration so we can easily get both of these values from there.
-            $locale = $app['config']['app.locale'];
-
-            if ($app['config']['app.phrase_in_context']) {
-                $trans = new PterodactylTranslator($loader, $locale);
-            } else {
-                $trans = new Translator($loader, $locale);
-            }
-
-            $trans->setFallback($app['config']['app.fallback_locale']);
-
-            return $trans;
-        });
+        $key = substr($key, strpos($key, '.') + 1);
+        return "{{__phrase_${key}__}}";
     }
-
-
 }
